Refactor Icons of Player and Volume
